[develop] Code refactoring: renamed controllers, better classified in packages
