+++ mapito/ProfileController.swift
@@ -2,64 +2,69 @@
 //  ProfileController.swift
 //  mapito
 //
-//  Created by m2sar on 21/04/2018.
+//  Created by m2sar on 12/04/2018.
 //  Copyright © 2018 code2geny. All rights reserved.
 //
 
-import Foundation
 import UIKit
 
-class ProfileController: UIViewController {
+class ProfileController: UIViewController, UITableViewDelegate, UITableViewDataSource {
     
-    
-    @IBOutlet weak var newValueTextField: UITextField!
-    @IBOutlet weak var confirmNewValueTextField: UITextField!
-    //path de la methode modification attributs
-    @IBInspectable var modifURL: String!
-    
+    //deconnexion
+    @IBAction func deconnexion(_ sender: UIButton) {
+        let appDelegate = UIApplication.shared.delegate as! AppDelegate;
+        appDelegate.authenticated = false;
+        
+        let defaults = UserDefaults.standard;
+        defaults.removeObject(forKey: "token");
+        defaults.synchronize();
+        
+        let storyboard: UIStoryboard = UIStoryboard(name: "Main", bundle: nil);
+        let navigationAuthentication = storyboard.instantiateViewController(withIdentifier: "NavigationAuthentication") as! UINavigationController;
+        self.present(navigationAuthentication, animated: true);
+    }
+    //var indexArray = ["Info personnelles","Info Connexion"]
+    var champs = ["Modifier Nom","Modifier Prénom","Modifier Email","Modifier Mot de passe"]
     override func viewDidLoad() {
         super.viewDidLoad()
         // Do any additional setup after loading the view, typically from a nib.
     }
-    
+
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
         // Dispose of any resources that can be recreated.
     }
+    /*
+    //nb de separation de label
+    func numberOfSections(in tableView: UITableView) -> Int {
+        return indexArray.count
+    }
+    //afficher info perso info connexion
+    func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
+        return indexArray[section]
+    }
+    */
     
-    //lorsque le user clique sur modifier
-    @IBAction func modificationValue(_ sender: Any) {
-        let newValue = newValueTextField.text;
-        let confirm = confirmNewValueTextField.text;
+    //nb de label par cell
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return champs.count
+    }
+    //initialisation de la tableView
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: "TableCell", for: indexPath) //as! TableCell
         
-        //verif champs vide
-        if((newValue?.isEmpty)! || (confirm?.isEmpty)!){
-            displayMessage(userMessage: "Veuillez remplir correctement tous les champs");
-            return;
-        }
-        //Si les champs ne sont pas vide, alors appel methode d' inscription
-        let baseUrl = URL(string: self.modifURL)!
-        /*
-         let query: [String : String] = [
-         "newValue" : newValue!,
-         "confirm" : confirm!
-         ]*/
-        let request = URLRequest(url: baseUrl)
-        let session = URLSession.shared.dataTask(with: request , completionHandler: { (data, response, error) in
-            if let jsonData = String(data: data!, encoding: .utf8) {
-                print(jsonData)
-            }
-            //si la valeur existe deja
-        })
-        session.resume()
+        cell.textLabel?.text = champs[indexPath.row]
+        
+        return cell
     }
     
-    //message info
-    func displayMessage(userMessage: String)
-    {
-        var myAlert = UIAlertController(title: "Attention", message: userMessage, preferredStyle: UIAlertControllerStyle.alert);
-        let Ok = UIAlertAction(title: "Ok", style: UIAlertActionStyle.default, handler:nil);
-        myAlert.addAction(Ok);
-        self.present(myAlert, animated: true, completion: nil);
+    //une fois cliquer sur un label on est redirigé vers la page DetailedProfileController
+    func tableView(_ tableView: UITableView, didDeselectRowAt indexPath: IndexPath) {
+        let mainStory: UIStoryboard = UIStoryboard(name: "Main", bundle: nil)
+        let deCV = mainStory.instantiateViewController(withIdentifier: "DetailedProfileController") as! DetailedProfileController
+        self.navigationController?.pushViewController(deCV, animated: true)
+        //self.present(deCV, animated: true);
     }
+    
 }
+
