--- mapito/TabBarController.swift
@@ -1,25 +1,0 @@
-//
-//  TabBarController.swift
-//  mapito
-//
-//  Created by m2sar on 21/04/2018.
-//  Copyright © 2018 code2geny. All rights reserved.
-//
-
-import Foundation
-import UIKit
-
-class TabBarController: UITabBarController {
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-    }
-    
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-    
-    
-}
